refactor(ui): 将前端重构为分层架构，业务逻辑与视图解耦 (close  [Feature]: 重构Rust后端，分模块，分层管理 Fixes #38)
