+++ src/features/chat/components/ChatMessage.vue
