+++ src/features/models/components/ModelItem.vue
