+++ src/shared/components/BackgroundEffect.vue
